+++ src/main/java/school/faang/user_service/filters/event/EventByOwnerFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filters.Event;
+package school.faang.user_service.filters.event;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.event.EventFilterDto;
@@ -7,14 +7,18 @@ import school.faang.user_service.entity.event.Event;
 import java.util.stream.Stream;
 
 @Component
-public class EventTitleFilter implements EventFilter {
+public class EventByOwnerFilter implements EventFilter {
     @Override
     public boolean isApplicable(EventFilterDto filters) {
-        return filters.getTitlePattern() != null && !filters.getTitlePattern().isEmpty();
+        return filters.getOwnerIdPattern() != null;
     }
 
     @Override
     public Stream<Event> apply(Stream<Event> events, EventFilterDto filters) {
-        return events.filter(event -> event.getTitle().contains(filters.getTitlePattern()));
+        {
+            events.filter(event ->
+                            (event.getOwner().getId()) == (filters.getOwnerIdPattern()));
+            return events;
+        }
     }
 }
