+++ src/main/java/school/faang/user_service/filters/event/EventFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filters.Event;
+package school.faang.user_service.filters.event;
 
 import school.faang.user_service.dto.event.EventFilterDto;
 import school.faang.user_service.entity.event.Event;
