+++ src/main/java/school/faang/user_service/filters/invitation/InvitationInvitedIdNameFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filters.Invitation;
+package school.faang.user_service.filters.invitation;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
@@ -7,14 +7,14 @@ import school.faang.user_service.entity.goal.GoalInvitation;
 import java.util.stream.Stream;
 
 @Component
-public class InvitationInvitedNameFilter implements InvitationFilter {
+public class InvitationInvitedIdNameFilter implements InvitationFilter {
     @Override
     public boolean isAcceptable(InvitationFilterDto goal) {
-        return !goal.getInvitedNamePattern().isEmpty() && !goal.getInvitedNamePattern().isBlank();
+        return goal.getInvitedId() != null;
     }
 
     @Override
     public Stream<GoalInvitation> apply(Stream<GoalInvitation> goal, InvitationFilterDto filters) {
-        return goal.filter(invintation -> invintation.getInvited().getUsername().matches(filters.getInvitedNamePattern()));
+        return goal.filter(invintation -> invintation.getInvited().getId() == (filters.getInvitedId()));
     }
 }
