+++ src/main/java/school/faang/user_service/filters/invitation/InvitationStatusFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filters.Invitation;
+package school.faang.user_service.filters.invitation;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
