+++ src/main/java/school/faang/user_service/service/event/EventService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.event.EventFilterDto;
 import school.faang.user_service.exception.ResourceNotFoundException;
-import school.faang.user_service.filters.Event.EventFilter;
+import school.faang.user_service.filters.event.EventFilter;
 import school.faang.user_service.mapper.EventMapper;
 import school.faang.user_service.dto.event.EventDto;
 import school.faang.user_service.entity.event.Event;
