+++ src/test/java/school/faang/user_service/service/filters/InvitationInvitedIdNameFilterTest.java
@@ -7,7 +7,7 @@ import org.mockito.Mockito;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.goal.GoalInvitation;
-import school.faang.user_service.filters.Invitation.InvitationInvitedIdNameFilter;
+import school.faang.user_service.filters.invitation.InvitationInvitedIdNameFilter;
 
 import java.util.List;
 import java.util.stream.Stream;
