+++ src/test/java/school/faang/user_service/service/filters/InvitationInviterNameFilterTest.java
@@ -7,7 +7,7 @@ import org.mockito.Mockito;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.goal.GoalInvitation;
-import school.faang.user_service.filters.Invitation.InvitationInviterNameFilter;
+import school.faang.user_service.filters.invitation.InvitationInviterNameFilter;
 
 import java.util.List;
 import java.util.stream.Stream;
