+++ src/test/java/school/faang/user_service/service/filters/InvitationStatusFilterTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
 import school.faang.user_service.entity.goal.GoalInvitation;
-import school.faang.user_service.filters.Invitation.InvitationStatusFilter;
+import school.faang.user_service.filters.invitation.InvitationStatusFilter;
 
 import java.util.List;
 import java.util.stream.Stream;
