+++ src/test/java/school/faang/user_service/service/service/GoalInvitationServiceImplTest.java
@@ -12,7 +12,7 @@ import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.goal.Goal;
 import school.faang.user_service.entity.goal.GoalInvitation;
-import school.faang.user_service.filters.Invitation.InvitationFilter;
+import school.faang.user_service.filters.invitation.InvitationFilter;
 import school.faang.user_service.mapper.GoalInvitationMapper;
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.goal.GoalInvitationRepository;
